Break TokenSemanticInfo out into it's own file
